test(management-auth): :white_check_mark: add pass tests of users, roles and preference notification

+++ src/contexts/auth/application/use-cases/role/delete-role.ts
@@ -1,4 +1,3 @@
-import { ForeignKeyConstraintError } from 'sequelize';
 import { RoleRepository } from '@src/contexts/auth/core/repository/role-repository';
 import { AppError } from '@src/contexts/shared/infrastructure/exception/AppError';
 
@@ -12,13 +11,12 @@ export class DeleteRole {
       throw new AppError('ROLE_NOT_FOUND', 404, 'Role not found', true);
     }
 
-    try {
-      await this.roleRepository.delete(id);
-    } catch (error) {
-      if (error instanceof ForeignKeyConstraintError)
-        throw new AppError('ROLE_HAS_RELATIONS', 400, 'You can not delete a role with active users', true);
+    const activeUsers = await this.roleRepository.getUsersActiveByRoleId(id);
 
-      throw error;
+    if (activeUsers.length > 0) {
+      throw new AppError('CANNOT_DELETE_ACTIVE_ROLE', 400, 'You can not delete a role with active users', true);
     }
+
+    await this.roleRepository.delete(id);
   }
 }
